MMT-3556: Changing jest to vi
